refactor: merge empty SwiftSync.swift stub into the namespace's own file
The Core.swift split left SwiftSync.swift holding only `public enum SwiftSync {}`
beside a SwiftSync+Helpers.swift whose name was the vague grab-bag suffix we
reject elsewhere. Those 11 statics aren't a foreign-type feature — they are
SwiftSync's own general-purpose methods, so they belong in the type's own file.

Declare `enum SwiftSync` with its statics directly in SwiftSync.swift; delete
the stub and the +Helpers name. Specialized SwiftSync extensions (macro SPI in
SyncableMacroSupport.swift, profiling in SyncPerformanceProfiler.swift) keep
their concern files — the legitimate namespace pattern.

Pure move: SwiftSync 172 (+30), DemoCore unaffected — green.
